Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,7 +2,7 @@
 
 
 void main() {
-    double a;
-    scanf("%u",&a);
-    printf("You entered: %u", a);
+    float a;
+    scanf("%f",&a);
+    printf("You entered: %f", a);
 }
